requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
contourpy==1.3.3
cycler==0.12.1
filelock==3.13.1
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2024.6.1
h5py==3.14.0
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.8
lightning==2.5.2
lightning-utilities==0.15.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.6.3
networkx==3.3
numpy==2.1.2
packaging==25.0
pillow==11.0.0
propcache==0.3.2
protobuf==6.31.1
ptwt==1.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.2
PyWavelets==1.8.0
PyYAML==6.0.2
scipy==1.16.1
six==1.17.0
sympy==1.13.3
tensorboardX==2.6.4
torch==2.7.1+cpu
torchaudio==2.7.1+cpu
torchmetrics==1.8.0
torchvision==0.22.1+cpu
tqdm==4.67.1
typing_extensions==4.12.2
yarl==1.20.1
